add test of module property
