+++ python/oneflow/test/graph/test_graph_block.py
@@ -330,6 +330,35 @@ class TestGraphBlock(flow.unittest.TestCase):
         # print(para_dict_g)
         test_case.assertTrue(np.array_equal(output_m.numpy(), output_g.numpy()))
 
+    def test_block_get_dtype_property(test_case):
+        class ModuleDtype(flow.nn.Module):
+            def __init__(self):
+                super().__init__()
+                self.linear = flow.nn.Linear(3, 8, False).to("cuda")
+                print(self.linear.weight.dtype)
+            
+            @property
+            def dtype(self):
+                return flow.float64
+            
+            def forward(self, x):
+                dt = self.dtype
+                print(dt)
+                self.linear
+                print(self.linear.weight.dtype)
+                return self.linear(x)
+        
+        m = ModuleDtype()
+
+        np_in = np.ones((8, 3)).astype(np.float32)
+        x = flow.tensor(np_in, device="cuda")
+
+        @flow.nn.Graph.to_graph
+        def block_get_dtype(x):
+            return m(x)
+        
+        out = block_get_dtype(x)
+
 
 if __name__ == "__main__":
     unittest.main()
